backend/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.4
djangorestframework==3.10.3
pyjwt==1.5.0
psycopg2-binary==2.8.3
pycparser==2.19
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
